# doc/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:9090",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/plc/writeMPoint": {
            "post": {
                "tags": [
                    "controller-plc"
                ],
                "operationId": "writeMPoint",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestWriteMPointDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/plc/writeDPoint": {
            "post": {
                "tags": [
                    "controller-plc"
                ],
                "operationId": "writeDPoint",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestWriteDPointDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/plc/MPointState": {
            "get": {
                "tags": [
                    "controller-plc"
                ],
                "operationId": "MpointState",
                "parameters": [
                    {
                        "name": "param",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/RequestMPointStateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/plc/DPointData": {
            "get": {
                "tags": [
                    "controller-plc"
                ],
                "operationId": "DPointData",
                "parameters": [
                    {
                        "name": "param",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/RequestDPointStateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RequestWriteMPointDTO": {
                "type": "object",
                "properties": {
                    "param": {
                        "type": "string"
                    },
                    "value": {
                        "type": "boolean"
                    }
                }
            },
            "RequestWriteDPointDTO": {
                "type": "object",
                "properties": {
                    "param": {
                        "type": "string"
                    },
                    "value": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "RequestMPointStateDTO": {
                "type": "object",
                "properties": {
                    "param": {
                        "type": "string"
                    }
                }
            },
            "RequestDPointStateDTO": {
                "type": "object",
                "properties": {
                    "param": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}